+++ src/components/HandlerManager/HandlerManager.js
@@ -2,14 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withModules } from '../Modules';
 import { stripesShape } from '../../Stripes';
-import withHandlers from './withHandlers';
 
-class Handlers extends React.Component {
+import { getHandlerComponents } from '../../handlerService';
+
+class HandlerManager extends React.Component {
   static propTypes = {
     stripes: stripesShape.isRequired,
     event: PropTypes.number,
     data: PropTypes.object,
-    getComponentsFromHandlers: PropTypes.func,
     modules: PropTypes.shape({
       handler: PropTypes.array,
     }),
@@ -17,8 +17,8 @@ class Handlers extends React.Component {
 
   constructor(props) {
     super(props);
-    const { event, stripes, modules, getComponentsFromHandlers, data } = props;
-    this.components = getComponentsFromHandlers(event, stripes, modules.handler, data);
+    const { event, stripes, modules, data } = props;
+    this.components = getHandlerComponents(event, stripes, modules.handler, data);
   }
 
   render() {
@@ -28,4 +28,4 @@ class Handlers extends React.Component {
   }
 }
 
-export default withModules(withHandlers(Handlers));
+export default withModules(HandlerManager);
