--- src/components/Handlers/index.js
@@ -1,7 +1,0 @@
-import withHandlers from './withHandlers';
-import Handlers from './Handlers';
-
-export {
-  withHandlers,
-  Handlers,
-};
