--- src/components/MainNav/MyProfile/index.js
@@ -1,1 +1,0 @@
-export { default } from './MyProfile';
